requirements.txt
Django==2.2.1
django-markdownx==2.0.28
Markdown==3.1
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
